refactor(shared): consolidate manifest-I/O plumbing
manifest_io.py byte-duplicated ManifestLoadError and parse_mapping from
easy_cheese_schemas.io, so `except ManifestLoadError` imported from one
module silently missed exceptions raised from the other. Re-export the
schemas package's copies instead of shadowing them locally.

Add manifest_io.json_command(func, usage) so age_route_cli,
pasteurize_route_cli, and press_route_cli stop hand-rolling the same
read-payload/call/print-JSON wrapper; exit codes (2/1/0) and the
"ERROR: " stderr prefix are preserved byte-exactly.

manifest_io now imports easy_cheese_schemas, so every .pyz bundle that
stages it picks up the schemas wheel transitively via easy-cheese-shared's
existing dependency; regenerated the bundle locks and rebuilt all 14
.pyz archives to match.
