+++ requirements/bundles/briesearch.txt
@@ -2,5 +2,5 @@ attrs==26.1.0 --hash=sha256:c647aa4a12dfbad9333ca4e71fe62ddc36f4e63b2d260a37a8b8
 cattrs==26.1.0 --hash=sha256:d1e0804c42639494d469d08d4f26d6b9de9b8ab26b446db7b5f8c2e97f7c3096
 easy-cheese-briesearch==1.1.0 --hash=sha256:3fb52d1e4049a819bd5cccfbda3f5752d8f7ddf1ccbcdc5d8eb6a95dcd5ecc8b
 easy-cheese-schemas==1.1.0 --hash=sha256:cf82f832dfed5143a64b2a884ed82a6e9d4e5fb4c317cfb3bb9a2ebc527598a0
-easy-cheese-shared==1.1.0 --hash=sha256:785b6516fc571d9d7b9fa0ecd312c9cd6be559967521d1eb600b263390c56cf5
+easy-cheese-shared==1.1.0 --hash=sha256:920ef8000a156e2782053bad9ba79f05c3746ade73a71c7be6c8e4827f4aa3d1
 typing_extensions==4.16.0 --hash=sha256:481caa481374e813c1b176ada14e97f1f67a4539ce9cfeb3f350d78d6370c2e8
