+++ src/easy_cheese/shared/fanout/press_route_cli.py
@@ -2,11 +2,10 @@
 
 from __future__ import annotations
 
-import json
 import sys
 from dataclasses import asdict
 
-from easy_cheese.shared.manifest_io import ManifestLoadError, read_mapping_arg_or_stdin
+from easy_cheese.shared.manifest_io import json_command
 
 from .press_route import (
     Continue,
@@ -15,6 +14,13 @@ from .press_route import (
     Stop,
     route_from_receipt,
 )
+
+_EXPECTED_KEYS = {
+    "outcome",
+    "current_receipt",
+    "phase_token_ref",
+    "phase_token_sha256",
+}
 
 
 def _action_payload(action: Continue | Dispatch | Stop) -> dict[str, object]:
@@ -27,39 +33,25 @@ def _action_payload(action: Continue | Dispatch | Stop) -> dict[str, object]:
     return {"action": name, **asdict(action)}
 
 
-def main(argv: list[str]) -> int:
-    try:
-        payload = read_mapping_arg_or_stdin(
-            argv,
-            "usage: press_route_cli.py [<request.json>]",
+def _route(**payload: object) -> dict[str, object]:
+    if set(payload) != _EXPECTED_KEYS:
+        raise ValueError(
+            "request must contain exactly outcome, current_receipt, "
+            "phase_token_ref, and phase_token_sha256"
         )
-    except ManifestLoadError as exc:
-        print(f"ERROR: {exc}", file=sys.stderr)
-        return 2
     try:
-        expected = {
-            "outcome",
-            "current_receipt",
-            "phase_token_ref",
-            "phase_token_sha256",
-        }
-        if set(payload) != expected:
-            raise ValueError(
-                "request must contain exactly outcome, current_receipt, "
-                "phase_token_ref, and phase_token_sha256"
-            )
         action = route_from_receipt(
             payload["outcome"],
             payload["current_receipt"],
             payload["phase_token_ref"],
             payload["phase_token_sha256"],
         )
-    except (ReceiptChainError, TypeError, ValueError) as exc:
-        print(f"ERROR: {exc}", file=sys.stderr)
-        return 1
-    json.dump(_action_payload(action), sys.stdout, indent=2)
-    sys.stdout.write("\n")
-    return 0
+    except ReceiptChainError as exc:
+        raise ValueError(str(exc)) from exc
+    return _action_payload(action)
+
+
+main = json_command(_route, "usage: press_route_cli.py [<request.json>]")
 
 
 if __name__ == "__main__":
