+++ src/easy_cheese/shared/manifest_io.py
@@ -5,35 +5,16 @@ from __future__ import annotations
 import json
 import sys
 from pathlib import Path
-from typing import Any
+from typing import Any, Callable
 
+from easy_cheese_schemas.io import ManifestLoadError, parse_mapping
 
-class ManifestLoadError(Exception):
-    """Raised when a manifest-like document cannot be loaded as a mapping."""
-
-
-def parse_mapping(text: str, source: str = "<stdin>") -> dict[str, Any]:
-    """Parse JSON first, then YAML, and require a top-level mapping."""
-    try:
-        data = json.loads(text)
-    except json.JSONDecodeError as json_exc:
-        try:
-            import yaml
-        except ImportError as exc:
-            raise ManifestLoadError(
-                f"{source}: invalid JSON and PyYAML is not installed for YAML input: {json_exc}"
-            ) from exc
-        try:
-            data = yaml.safe_load(text)
-        except yaml.YAMLError as yaml_exc:
-            raise ManifestLoadError(
-                f"{source}: invalid JSON ({json_exc}) and invalid YAML ({yaml_exc})"
-            ) from yaml_exc
-
-    if not isinstance(data, dict):
-        raise ManifestLoadError(f"{source}: expected a mapping at document root")
-    return data
-
+__all__ = [
+    "ManifestLoadError",
+    "parse_mapping",
+    "read_mapping_arg_or_stdin",
+    "json_command",
+]
 
 def read_mapping_arg_or_stdin(argv: list[str], usage: str) -> dict[str, Any]:
     """Read one optional path argument or stdin, returning a parsed mapping."""
@@ -47,3 +28,28 @@ def read_mapping_arg_or_stdin(argv: list[str], usage: str) -> dict[str, Any]:
             raise ManifestLoadError(f"manifest not found: {path}") from exc
         return parse_mapping(text, str(path))
     return parse_mapping(sys.stdin.read())
+
+
+def json_command(func: Callable[..., Any], usage: str) -> Callable[[list[str]], int]:
+    """Build a `main(argv)` that reads a JSON mapping, calls `func`, and prints JSON.
+
+    Exit codes and the "ERROR: " stderr prefix match the previously hand-written
+    wrappers: 2 for a manifest-load failure, 1 for a `func` rejection, 0 on success.
+    """
+
+    def main(argv: list[str]) -> int:
+        try:
+            payload = read_mapping_arg_or_stdin(argv, usage)
+        except ManifestLoadError as exc:
+            print(f"ERROR: {exc}", file=sys.stderr)
+            return 2
+        try:
+            result = func(**payload)
+        except (TypeError, ValueError) as exc:
+            print(f"ERROR: {exc}", file=sys.stderr)
+            return 1
+        json.dump(result, sys.stdout, indent=2)
+        sys.stdout.write("\n")
+        return 0
+
+    return main
